Added new versions of scripts.
Changed the data structure within `generate_all_matrices` to be indices of non-zero elements instead of `numpy` arrays. This gives the same run time while give a large improvement in the use of memory. To allow users to run the old versions those functions have been left in and a boolean flag `element_run` which allows you to toggle between the two data structures.