Change data storage for matrices, added lattice drawing
The way that matrices are stored in data now are that each row is converted from [0,1,0,0,0,0,0,...] to a binary number using np.packbit. This reduces the memory cost of storing the arrays from O(k^2) to O(k).

drawTrangleLattice.py : given an interaction matrix this program produces a PNG file that represents the deterministic tiling of the pattern. Adapted from Daichi's code to use a `class` structure for each triangle. Improves readability and might be useful for identifying primitive vectors of the tiling.