Searches tree branches in full
This restructures the way the code runs. Instead of generating each level of the tree structure and storing every possible matrix till the end, this searches down a branch to the root. This generates a full matrix that we can check if its valid. If valid it is stored and all degenerate matrices are made to provide additional checks against.

Right now the generation of all degenerate matrices is extremely slow since it has to create O(3(k!)) from the original.